Домашнее задание 7. Рефакторинг, правки по замечаниям ПР
