+++ library-spring-data/src/main/java/ru/otus/library/facade/AuthorFacade.java
@@ -1,0 +1,13 @@
+package ru.otus.library.facade;
+
+public interface AuthorFacade {
+
+    void addAuthor();
+
+    void findAuthorByName(String authorName);
+
+    void findAllAuthors();
+
+    void deleteAuthor(Long id);
+
+}
