+++ library-spring-data/src/main/java/ru/otus/library/facade/BookFacade.java
@@ -1,0 +1,17 @@
+package ru.otus.library.facade;
+
+public interface BookFacade {
+
+    void addBook();
+
+    void findBookByName(String name);
+
+    void findBooksByAuthor(String authorName);
+
+    void findAllBooks();
+
+    void deleteBook(Long id);
+
+    void updateBook(Long id);
+
+}
