+++ library-spring-data/src/main/java/ru/otus/library/facade/GenreFacade.java
@@ -1,0 +1,13 @@
+package ru.otus.library.facade;
+
+public interface GenreFacade {
+
+    void addGenre();
+
+    void findGenreByName(String name);
+
+    void findAllGenre();
+
+    void deleteGenre(Long id);
+
+}
