+++ library-spring-data/src/main/java/ru/otus/library/repository/BookRepository.java
@@ -11,7 +11,4 @@ public interface BookRepository extends CrudRepository<Book, Long> {
 
     List<Book> findByName(String name);
 
-    List<Book> findByAuthor_Name(String name);
-
-    List<Book> findByGenre_Name(String name);
 }
